Reintroduce the GCC plugin
It should be possible to build the GCC plugin both in tree, and out of
tree. This adds the ability to include the plugin in the build if it is
available in the gcc_plugin directory.

Some minor refactoring has been done to the framework, and helpful
print methods for features/parameters and their sets have been defined.

ChangeLog:

	* CMakeLists.txt: Refactor, build libraries and add option to build
	plguin.
	* include/mageec/Feature.h (FeatureBase::print)
	(Feature::print): Added methods.
	* include/mageec/FeatureSet.h (FeatureSet::addFeature): rename to add.
	(FeatureSet::print): Added method.
	* include/mageec/Parameter.h (ParameterBase::print)
	(Parameter::print): Added methods.
	* include/mageec/ParameterSet.h (ParameterSet:addParameter): rename to add.
	(ParameterSet::print): Added method.
	* lib/Database.cpp (ResultIterator::operator*): Update addFeature/
	addParameter calls.