Minor refactor of some interfaces, fixed Option type
The Option type assignment operator was incorrect, and the LHS
of the assignment would be left in an erroneous state.

UUIDs can now be parsed from strings in the UUID canonical form. The
loaded UUIDs are stored big endian form internally.

TrainedMLs can be queried to retrieve their UUID. This is necessary in
order to allow the user to use a UUID to select a trained machine learner out
of those returned from the database. A UUID is also returned when a machine
learner is loaded from a plugin, so that the user can identify it among
the trained machine learners.

ChangeLog:

	* include/mageec/Framework.h (Framework::loadMachineLearner):
	Return the UUID of the loaded machine learner.
	* include/mageec/TrainedML.h (TrainedML::getUUID): Added function.
	* include/mageec/Util.h (UUID): Added default copy/move constructors
	and assignment.
	(UUID::parse): Added method to parse a UUID from a canonical string.
	(Option): Fix copy/move constructors and assignment operators.
	* lib/Framework.cpp (Framework::loadMachineLearner): Return UUID.
	* lib/TrainedML.cpp (TrainedML::getUUID): Added implementation.