Refactored to use hashes for some id fields in the database
Previously, we used a monotonically increasing number to identify each
pass sequence, feature set, feature group and parameter set. This meant
that duplicate features sets could not be identified, which is useful
for some machine learners, and for avoiding bloat.

Now a hash is used instead so that a certain table entries can be
identified by hashing their constituent features. This means that when
inserting new entries into the database, we avoid inserting
duplicate entries if possible.

Before, there were a number of tables which existed solely to keep track
of the next available id. These have now been removed as they are
unnecessary. As a consequence of this, the database can no longer
enforce referential integrity for some ids using foreign keys, and this
has to be enforced externally.

Finally, a couple of functions have had their interfaces simplified, as
feature sets, parameter sets, and pass sequences can no longer be
constructed incrementally.

ChangeLog:

	* include/Database.h (Database::newFeatureGroup): Refine to take a
	set of FeatureSetID rather than vector.
	(Database::addPasses): Removed method.
	(Database::newPassSequence): Method now takes a list of passes to
	add to the database.
	* include/Types.h (CompilationID FeatureSetID FeatureGroupID)
	(PassID PassSequenceID ParameterSetID ParameterGroupID): Redefined
	types to be uint64_t so that they can store a hash.
	* include/Util.h (read64LE wrte64LE): Added functions.
	* lib/Database.cpp (create_feature_set_table)
	(create_feature_group_table create_parameter_set_table)
	(create_pass_sequence_table): Remove query strings to create
	some tables.
	(Database::init_db): Remove creation of some tables.
	(Database::newFeatureSet): The returned FeatureSetID is now a hash of
	the input features, the database no longer contains duplicate feature
	sets.
	(Database::newFeatureGroup): Returned FeatureGroupID is now a hash,
	the database no longer contains duplicate feature groups.
	(Database::newParameterSet): The returned ParameterSetID is now a
	hash of the input features, the database no longer contains duplicate
	parameter sets.
	(Database::newPassSequence): The returned PassSequenceID is now a
	hash of the input passes. The database no longer contains duplicate
	pass sequences.
	* lib/Main.cpp (main): Updated to use new interfaces.
	* lib/Util.cpp (read64LE write64LE): Added function implementations.