Add FileML and update machine learner interfaces.
FileML is a non-learner machine learner. It makes decisions based on
predefined and user provided list. This means that it can be used to
drive the compilation configuration through the MAGEEC plugin rather
than by the user manually specifying the configuration when invoking
the compiler.

This requires some changes to the machine learner interface. Importantly
it means that we have a machine learner which does not need to be
'trained', and which doesn't store anything in the database. It also
requires that we must have some method for the user to specify the
desired compiler configuration.

Machine learners may now specify that they don't need to be trained,
and there is also the ability to pass a configuration file through to
the machine learner both when training and making decisions.

Currently, the new capabilities are not exposed through the gcc plugin
or the standalone tool.

ChangeLog:

	* CMakeLists.txt: Add FileML.cpp to ML_SOURCES
	* include/mageec/ML.h (IMachineLearner::requiresTraining)
	(IMachineLearner::requiresTrainingConfig)
	(IMachineLearner::setTrainingConfig)
	(IMachineLearner::requiresDecisionConfig)
	(IMachineLearner::setDecisionConfig): Added methods.
	* include/mageec/ML/C5.h (C5Driver::requiresTraining)
	(C5Driver::requiresTrainingConfig C5Driver::setTrainingConfig)
	(C5Driver::requiresDecisionConfig C5Driver::setDecisionConfig):
	Likewise.
	* include/mageec/ML/FileML.h (FileML): Added definition of FileML.
	* include/mageec/TrainedML.h (TrainedML::requiresDecisionConfig)
	(TrainedML::setDecisionConfig): Added methods.
	* lib/Database.cpp (getTrainedMachineLearners): Make IMachineLearner
	non-const.
	* lib/FileML.cpp (FileML): Added implementation of FileML.
	* lib/TrainedML.cpp (TrainedML::TrainedML): Make IMachineLearner
	parameter non-const.
	(TrainedML::requiresDecisionConfig)
	(TrainedML::setDecisionConfig): Added implementations.