Refactor interfaces, allow machine learners which don't need training
Before, machine learners were only considered trained if they had a
blob stored in the database. With FileML it was necessary to allow a
machine learner which did not need a training phase, and therefore
a number of interfaces needed to be updated.

A consequence of these changes is that a machine learner does not need
to store anything in the database. If desired, the machine learner could
store its training data externally, and require the user to provide it
through the configuration interface. The only limitation is that it
would still need to use the training interface, as there is no other
way to retrieve the results from the database.

The interfaces are still quite clumsy, and need further work. For
example, the 'getTrainedMachineLearners' function in Database.cpp can
only return machine learners which exist in the database. So the
function does not return TrainedMLs for machine learners which did not
need training in the first place.

ChangeLog:

	* include/mageec/TrainedML.h (TrainedML::TrainedML): Added constructor
	which does not take a metric or blob argument.
	(TrainedML::TrainedML): Removed db argument from constructors as it
	was unnecessary.
	(TrainedML::m_db): Removed member.
	* include/mageec/Util.h (UUID::operator!=): Added method.
	(metricToString stringToMetric): Moved functions from Driver.cpp so
	that they are accessible to the plugin.
	* lib/Database.cpp (Database::getTrainedMachineLearners): Update
	creation of TrainedML.
	* lib/Driver.cpp (metricToString stringToMetric): Moved functions
	to Util.cpp.
	(printTrainedMLs): Update so that db_path is optional. Return a
	bool result instead of int.
	(createDatabase): Return a bool instead of int result.
	(trainDatabase): Likewise.
	(main): Minor refactor of variable names.
	* lib/TrainedML.cpp (TrainedML::TrainedML): Added constructor
	implementation.
	(TrainedML::TrainedML): Remove db argument from constructors.
	* lib/Util.cpp (metricToString stringToMetric): Moved implementation
	from Driver.cpp.