Link against machine learners, refactor argument parsing.
Before, we did not link with the machine learners, or load them when
the framework was constructed.

The interface to the plugin has been simplified and a couple of new
features added.

The -print-ml-interfaces and -print-trained-mls arguments currently
do nothing. Likewise there is not yet a 'help' interface for the
plugin.

gcc_plugin/ChangeLog:

	* CMakeLists.txt: Link the plugin library with the machine learners.
	* FeatureExtract.cpp (featureExtractExecute): Update uses of the
	debug interface.
	* GCCCallBacks.cpp: Likewise.
	(mageecStartFile): Update to use new MAGEECContext structure.
	(mageecPassGate): Likewise.
	* GCCPlugin.cpp (parseArguments): Refactor argument parsing. Add
	debug, ml-config, metric and print-trained-mls arguments. Remove
	the old 'mode' argument in favor of seperate optimize and save
	features flags.
	(plugin_init): Initialize the built in C5 and FileML machine learners.
	* MAGEECPlugin.h (MAGEECContext): Redefine the context structure.