Fix various issues with the C5.0 machine learner
The C5.0 machine learner can only train for the best result for a given
input feature set, so this information must be extracted from the full
set of results.

The machine learner was also only training for parameters, not passes,
so this functionality has been added.

There were various bugs and errors introduced in the original
implementation which have been fixed.

ChangeLog:

	* include/mageec/FeatureSet.h (FeatureSet::size): Added method.
	* include/mageec/Result.h (Result::getPassName): Removed method.
	(Result::getPassSequence): Added method.
	* lib/Database.cpp (Database::trainMachineLearner): Extract
	the pass sequence from the database as well as the parameters and
	features.
	(ResultIterator::operator*): Correct select_features query to
	extract the correct identifiers.
	* lib/ML/C5.cpp (ResultComparator): Removed comparator object.
	(ResultFeatureComparator): Added comparator object to compare
	feature sets.
	(C5Driver::train): Use only the best results for each distinct
	feature set. Add training for passes too.
	* lib/ML/C5/CMakeLists.txt: Install the c5.0 machine learner binary.