Remove iterative training interface
This interface was intended to allow a machine learner to train using
new training data, plus a previously trained blob. Although it is a
desirable feature to have, and may be necesasry for future machine
learners, it is not needed right now.

When the need arises, this interface will be reimplemented.

ChangeLog:

	* include/mageec/ML.h: Remove iterative training interfaces.
	* include/mageec/ML/C5.h: Likewise.
	* include/mageec/ML/FileML.h: Likewise.
	* lib/ML/C5.cpp: Likewise.
	* lib/ML/FileML.cpp: Likewise.