Remove continuous feature extraction support
The role of continuous feature extraction was to run the feature
extractor after each optimization pass. This mean a greater amount of
training data would be provided to the machine learner.

This has been removed as it is not yet necessary, and adds complexity
to the database. It is envisioned that it will be readded at a later
date.

ChangeLog:

	* include/mageec/Database.h (Database::addFeaturesAfterPass):
	Removed method, continuous feature extraction interface removed.
	* lib/Database.cpp (create_compilation_feature_table):
	Removed SQL query.