Emit the filename into the output from the plugin
We need to be able to associate results with the functions and module
compilation ids, however to do this we must be able to identify the
source file which those function and modules ids correspond to.

If we have the source file, then we can match against the executable
files by reading entries in the DWARF debug information.

gcc_plugin/ChangeLog:

	* GCCCallbacks.cpp (mageecFinishFile): Emit the full filename into the
	compilation id file.
	* GCCPlugin.cpp (parseArguments): Open the output file for appending
	instead of writing.