Add scripts to generate and measure benchmarks
The benchmark build script generates and builds a benchmark with gcc and
the mageec plugin. A benchmark can be anything with an autoconf or cmake
build system, and it should work provided that the benchmark can be
built and installed out of tree. The benchmark must allow arbitrary
CFLAGS to be specified when building.

There are some constraints on the structure of the benchmarks.
Importantly if the benchmark builds multiple executables as output, then
those executables cannot have any common source files. This limitation
exists because there needs to be a one to one mapping between the source
file and a compilation id.

ChangeLog:

	* CMakeLists.txt: Install newly added scripts.
	* scripts/build-and-measure.py: New file.
	* scripts/generate-build.py: Likewise.
	* scripts/measure-code-size.py: Likewise.