Remove the original GCC plugin
The old gcc plugin tried to do both feature extract and pass gating.
However, doing both in the same plugin was unnecessary and made the
interface overcomplicated.

For now, this is being replaced with a much simpler plugin which only
does feature extraction. Once this is working, we can readd the ability
to make pass decisions within the compiler. For now, we will be limited
to making decisions on compilation flags at the module level. Once this
has been proven to work more complexity will be added back in.

ChangeLog:

	* CMakeLists.txt: Remove gcc_plugin from the build system.
	* gcc_plugin/.gitignore: Deleted file.
	* gcc_plugin/AUTHORS: Likewise.
	* gcc_plugin/CMakeLists.txt: Likewise.
	* gcc_plugin/COPYING: Likewise.
	* gcc_plugin/ChangeLog: Likewise.
	* gcc_plugin/FeatureExtract.cpp: Likewise.
	* gcc_plugin/Features.h: Likewise.
	* gcc_plugin/GCCCallbacks.cpp: Likewise.
	* gcc_plugin/GCCPlugin.cpp: Likewise.
	* gcc_plugin/MAGEECPlugin.h: Likewise.
	* gcc_plugin/Parameters.h: Likewise.
	* gcc_plugin/README: Likewise.