Add GCC wrapper driver for MAGEEC
The purpose of this driver is to wrap around gcc, so that the MAGEEC
framework can control the optimization flags which are provided to GCC.
The driver works entirely externally to gcc, and builds a command-line
which can be spawned to do the compilation itself.

There are two modes which the driver can operate in. 'gather' mode is
where the user is controlling the flags which are being built with, and
the purpose of the driver is to record the compilation in the database,
along with the feature (previously extracted) and the parameters used to
build the given program.

The other mode is 'optimize' mode. Here the user provides a machine
learner to use, and a metric to train again. The framework makes a
decision based on previously extracted features about the most
desireable parameters, and these are used to do the compilation.

Currently, this driver is untested and barebones. The 'optimize' and
'gather' operations are not yet functional.

tools/gcc_driver/ChangeLog:

	* AUTHORS: Added file.
	* CMakeLists.txt: Likewise.
	* COPYING: Likewise.
	* Driver.cpp: Likewise.
	* Parameters.h: Likewise.
	* README: Likewise.