Generate compilations for functions as well as modules
This allows results to be associated with the functions in the module.
Functions have a feature group generated by the feature extractor, and
the parameters used in the compilation are inherited from the
compilation of the module they are contained within.

Generating compilation ids for functions is also necessary in order to
collate results for the module, as we need to know the functions results
to sum to build the result for the module.

tools/gcc_driver/ChangeLog:

	* Driver.cpp: Generate compilation ids for functions as well as
	modules.