Remove FeatureGroupSet from the database, replace with feature classes
Before, we did not distinguish in the framework between module level
features and function level features. This was instead handled by the
feature extract which just made sure to assign distinct feature
identifiers to function and module features.

The hope was that the machine learner would be able to learn to
distinguish between function and module features. So in the interface to
the machine learner we would just provide a flat set of features to make
decisions against.

In reality this actually makes everything more complicated, and it may
be asking too much of the machine learner. As a consequence, this change
adds a mechanism to distinguish between different sets of features with
an extra identifier called the feature class. One feature class is used
for functions, and one for modules, and the machine learner is trained
against each of them independently. This has made the feature groups in
the database redundant, so they have been removed.

ChangeLog:

	* include/mageec/Database.h (Database::newFeatureGroup)
	(Database::getFeatureGroupFeatures): Remove functions.
	(Database::newCompilation): Add features_class parameter.
	(Database::trainMachineLearner): Add feature_class parameter.
	(ResultIterator::ResultIterator): Likewise.
	(ResultIterator::m_metric): Remove unused member.
	* include/mageec/TrainedML.h (TrainedML::TrainedML): Add
	feature_class parameter.
	(TrainedML::getFeatureClass): Add method.
	(TrainedML::m_feature_class): Add private member.
	* include/mageec/Types.h (FeatureGroupID): Remove enumeration.
	(FeatureClass): Add enumeration.
	* lib/Database.cpp (create_feature_group_set_table): Remove global.
	(create_compilation_table): Update sql to add feature_class_id column.
	(create_machine_learner_table): Likewise.
	(create_program_unit_debug_table): Rename to
	create_compilation_debug_table.
	(Database::init_db): Remove create_feature_group_set_table query.
	(Database::getTrainedMachineLearners): Retrieve feature_class_id
	in sql query, use when creating new TrainedMLs.
	(Database::garbageCollect): Remove deletion of unused feature groups.
	(Database::newFeatureGrouo): Remove function.
	(Database::getFeatureGroupFeatures): Likewise.
	(Database::newCompilation Database::trainMachineLearners):
	Add feature_class parameter.
	(Database::trainMachineLearner): Add feature_class_id to sql query
	adding a new MachineLearner.
	(ResultIterator::ResultIterator): Add feature_class parameter.
	(ResultIterator::m_metric): Remove member.
	* lib/Driver.cpp (trainDatabase): Also train against each feature
	class.
	* lib/TrainedML.cpp (TrainedML::m_feature_class): Add private member.
	(TrainedML::TrainedML): Add feature_class parameter.
	(TrainedML::getFeatureClass): Added function.