Make gcc plugin work with gcc versions >=4.8.0
The plugin interface was introduced in 4.5.0, and has been changed over
time. This adds #ifdefs to ensure it builds with versions of GCC over
4.8.0.

It should be possible to build the plugin for 4.5.0 as well, however the
tree.h header in the gcc plugin interface contains a structure with a
member called thread_local. 'thread_local' was introduced as a builtin
type in C++11, so this causes an error in the frontend.

plugin/gcc_feature_extract/ChangeLog:

	* FeatureExtract.cpp: Fix to work with all GCC versions >=4.8.0
	* Plugin.cpp: Likewise.