Relax check for duplicate feature entries in features file
The file containing identifiers for each unit of the program is read
in by the driver. This file contains a number of key value entries,
where the key is a string identifying a module or function in the
program, and the value is a number identifying the features for that
module or function in the database.

Sometimes this file can contain duplicate entries, for example if a
source file was run through the feature extractor multiple times.
If the duplicate entries do not match, then it may signal a problem and
a warning is emitted. The check which triggered the emission of this
warning was wrong, and so a lot of false positives were being reported.

tools/gcc_driver/ChangeLog:

	* Driver.cpp: Compare feature file entries on the name of the
	program unit. Relax an overly aggressive warning to remove
	false positives.