Add ability to call gfortran as well as gcc from the wrapper driver
The wrapper defers to a provided compiler to do the actual
compilation, however a different compiler must be used depending
on whether fortran or c is being compiled. If the wrapper is
called through mageec-gcc, then it will defer to the c compiler
provided, and if it is called through mageec-gfortran, then it
will defer to the provided fortran compiler.

tools/gcc_driver/ChangeLog:

	* CMakeLists.txt: Add a symlink from mageec-gcc to
	mageec-fortran at install time.
	* Driver.cpp: Add the ability to specific a command to invoke
	gfortran from the GCC wrapper. When the driver is used, it will
	use the provided fortran compiler or c compiler depending on
	whether the driver is called through mageec-gcc or
	mageec-gfortran.
	(main): Refactor the argument handling to cope with fortran file,
	and input file provided mid-way through the command line string.