In the feature extract file, use absolute paths for filenames
The feature extractor emits an output file which maps from a source file
+ module/function name to a set of features in the database. Originally
the 'source file' would be whatever GCC uses as main_input_filename,
which may or may not be an absolute path. We now canonicalized this to
be the absolute path of the file.

In addition, in the features file the module name was originally the
same as the filename, however now the module name is canonicalized to
the base name of main_input_filename.

plugin/gcc_feature_extract/ChangeLog:

	* Plugin.cpp: In the output features, use an absolute path for the
	filename and the basename of the module as the module name.