Use absolute filenames internally in the gcc wrapper
In order for the gcc driver to function it must read the features
csv file emitted by the feature extractor. For each input file the
driver encounters, it finds the features for that file by using the
filename as a key.

The feature extractor canonicalizes the filenames in the features
csv file so that they are absolute filenames, so the gcc driver
must do the same.

tools/gcc_driver/ChangeLog:

	* Driver.cpp: Convert input filenames to be absolute filenames for
	internal use.