Add the ability to call g++ from the wrapper driver
The wrapper defers to a provided compiler to do the actual
compilation. A common wrapper is use for gcc, g++ and gfortran.
The wrapper has arguments -fmageec-gcc, -fmageec-g++, -fmageec-gfortran
to specify the compilers which should be invoked. The actual
compiler which is invoked depends on whether the wrapper was invoked
as mageec-gcc, mageec-g++ or mageec-gfortran.

tools/gcc_driver/ChangeLog:

	* CMakeLists.txt: Add a symlink from mageec-gcc to mageec-g++ at
	install time.
	* Driver.cpp: Add the ability to specify a command to invokve g++
	from the GCC wrapper. Depending on whether the mageec-gcc,
	mageec-gfortran or mageec-g++ is called this will use the
	corresponding command -fmageec-g?? command from the command line.