Use mangled names in the feature extractor
The symbols names which appear in the final executable are
mangled, but the function names used in the feature extractor were
unmangled. This meant it was impossible to match features back
to function names in the executable which meant the framework
could not work with C++ programs.

plugin/gcc_feature_extract/ChangeLog:

	* Plugin.cpp: Use a mangled name for functions in the feature
	extractor.