Update format of 'results' file output by measurement script
Previously this file had the simple format:

  compilation_id,metric,value

However, this has been updated to add more information about the
compilation which the compilation_id field refers to. The format of
each line is now as follows:

  source_file,compilation_type,compilation_name,"result",compilation_id,
  metric, value

This gives tools processing these results more information, and it
also more closely follows the formats of other csv files produced
by the framework.

ChangeLog:

	* script/measure-code-size.py: Update result file csv output to
	include more metadata about the compiled element