Differentiate between Function and Module features in the plugin
It is useful to distinguish between function and module level
features as the description of features in the database is associated
with the enumeration value for the feature. If a function and module
share the same enumeration value, then it does not affect the
operation of MAGEEC, but it means that the description string is
only stored for one of the features in the database.

plugins/gcc_feature_extract/ChangeLog:

	* Features.h: Distinquish between FunctionFeature and ModuleFeatures
	again, to help with debug.