Set SQLite to use an in-memory journal file
For now we use a memory rollback journal, as it improves performance when
we have lots of small transactions and short-lived journals. This WILL
corrupt the database if we crash mid transaction. If this proves to
be a problem in future, then we may want to move to something like
PERSIST instead, which is higher performance that the default (DELETE),
and also preserves safety.

ChangeLog:

	* lib/Database: Set the journaling mode on the SQLite database
	to be an in-memory journal.