Allow incomplete results during combined elimination
Before, we checked that each iteration of combined elimination
had results for a consistent set of functions and modules which
made up the program. This worked when benchmarking code size, as
there would always been a size results for every single in the file.

When measuring time/energy this doesn't work, as due to inlining
and inconsistency during measuring some functions may not have
associated results. Therefore, this check has been disabled, and
the results will have to be sanity checked manually.

ChangeLog:

	* scripts/combined-elimination.py: Allow incomplete results
	when measuring benchmarks.