Parallelize testing of flags in combine elimination
For combined elimination we need to test each flag in turn to
check whether it has any effect on the output result. Before
we tested each flag in sequence, however it is possible to
do all of these builds and runs in parallel.

We use a python process pool in order to spawn multiple builds of
the benchmark simultaneously.

At the moment this won't quite work, as although we can run
multiple builds simultaneously if each build is building in the
source tree, they will clobber each others files. To avoid this
we must create a copy of the source tree each time.

ChangeLog:

	* scripts/combined-elimination.py: Do multiple flag tests
	during combined elimination in parallel, using a process
	pool.