When building benchmarks build in a shadow copy of the source directory
If the benchmark is being built in tree by multiple processes
simultaneously, then each build may clobber the output files of
the others.

To avoid this, for each build we create a directory of symlinks to
the original source and then build in that directory instead. Each
build will then produce output files in a different directory to
the source tree.

At the moment, we use lndir to create the shadow directory. This
could probably be done in python instead to make it more portable
(and do a full directory copy if symlinks are not available)

ChangeLog:

	* script/mageec.py: Use lndir to create a shadow copy of a
	directory with symlinks to the original source.