Use realpath instead of abspath to get identifier of source files in an executable.
Originally, when measuring the size of an executable, we used os.abspath
to get an absolute path to the source files referred to in the DWARF
debug information. In the framework itself, we use 'realpath' on all
paths to make them absolute. However, whereas 'realpath' resolves
symlinks, 'os.abspath' does not, so we ended up with a paths of an
inconsistent form being used as identifiers.

Now we use realpath in the python as well so that the identifiers
are consistent.

ChangeLog:

	* script/measure-code-size.py: Use realpath to get the
	path of source files in the executable instead of abspath.