Avoid a race hazard adding features/params to the database
When adding features and parameters to the database, we have to
use the database table as a hash table. When probing the hash
table we did not ensure that multiple processes could not insert
into the database whilst probing. Now, before inserting features
and parameters into the database, we require an exclusive lock to
be acquired for the database.

ChangeLog:

	* lib/Database.cpp (Database::newFeatureSet): Fix implementation
	to avoid a race hazard when adding features to the database.
	(Database::newParameterSet): Avoid race hazard when adding
	parameters to the database.