Perform all builds single threaded
If a build is done in parallel then we run into some issues. Multiple
executions of the compiler can end up trying to write to the same
files, and we have to ensure that two compilations don't conflict
when writing to the database.

To avoid these issues, for now we just force all builds to be
single threaded.

ChangeLog:

	* script/combined-elimination.py (build_and_measure): Remove
	jobs argument.
	(combined_elimination): Likewise.
	* script/feature-extract.py: Remove --jobs flags, run
	feature extraction single threaded.
	* script/generate-builds.py (generate_configurations): Remove
	jobs argument.
	* script/mageec.py (build): Remove jobs argument, build
	benchmarks single threaded.
	(feature_extract): Likewise.