Update doxygen comments and doxygen configuration
This updates the doxygen documentation in the majority of
source files to make it more consistent. It also updates
the doxygen configuration file to fix various issues with
the output documentation.

The main problem with the doxygen configuration was that
it was catching all files in the source tree as inputs to
doxygen. So if the project was installed into the main
directory there would be duplicate copies of the headers
included into the documentation; One in the include/
directory; and one in the install directory of the project.

ChangeLog:

	* doc/Doxyfile.in: Update doxygen configuration to fix
	various issues with the generated documentation.
	* include/mageec/Attribute.h: Update doxygen comments.
	* include/mageec/AttributeSet.h: Likewise.
	* include/mageec/Database.h: Likewise.
	* include/mageec/Decision.h: Likewise.
	* include/mageec/Framework.h: Likewise.
	* include/mageec/ML.h: Likewise.
	* include/mageec/ML/1NN.h: Likewise.
	* include/mageec/ML/C5.h: Likewise.
	* include/mageec/Result.h: Likewise.
	* include/mageec/SQLQuery.h: Likewise.
	* include/mageec/TrainedML.h: Likewise.
	* include/mageec/Types.h: Likewise.
	* include/mageec/Util.h: Likewise.
	* lib/Driver.cpp: Likewise.
	* lib/ML/C5.cpp: Likewise.
	* lib/Util.cpp: Likewise.